# Parallelise duckdb resulting in e.g. 2-4x speedup on 6 core machine