--- a/tests/test_correctness_of_convergence.py
+++ b/tests/test_correctness_of_convergence.py
@@ -68,7 +68,7 @@ def test_splink_converges_to_known_params():
     # CREATE TABLE __splink__df_comparison_vectors_abc123
     # and modify the following line to include the value of the hash (abc123 above)
 
-    cvv_hashed_tablename = "__splink__df_comparison_vectors_f9bd31158"
+    cvv_hashed_tablename = "__splink__df_comparison_vectors_3f3fea0c5"
     linker.register_table(df, cvv_hashed_tablename)
 
     em_training_session = EMTrainingSession(