See [here](https://github.com/moj-analytical-services/splink/discussions/1830) for investigation into best salting settings

At the moment, DuckDB paraellelises little if any of the Splink workflow.  This PR causes DuckDB to fully parallelise salted workloads, meaning operations like `predict()` are multiple times faster (especially on machines with many cores)

See https://github.com/duckdb/duckdb/discussions/9710

<details>
<summary>
time it example 
</summary>

```python
import time

from splink.datasets import splink_datasets
from splink.duckdb.blocking_rule_library import block_on
from splink.duckdb.comparison_library import (
    exact_match,
    jaro_winkler_at_thresholds,
    levenshtein_at_thresholds,
)
from splink.duckdb.duckdb_comparison_level_library import (
    else_level,
    null_level,
)
from splink.duckdb.duckdb_linker import DuckDBLinker
from splink.duckdb.linker import DuckDBLinker

df = splink_datasets.historical_50k


df = df.drop("cluster", axis=1)

settings_dict = {
    "probability_two_random_records_match": 0.0001,
    "link_type": "dedupe_only",
}

linker = DuckDBLinker(df, settings_dict)
brs = linker._find_blocking_rules_below_threshold(1e7)
list(brs.sort_values("comparison_count", ascending=True)["blocking_columns_sanitised"])

br = block_on("first_name")


def get_brs(salt):
    brs = [
        ["gender", "occupation", "full_name"],
        ["gender", "occupation", "postcode_fake"],
        ["occupation", "full_name"],
        ["occupation", "postcode_fake"],
        ["gender", "occupation", "first_and_surname"],
        ["first_name", "gender", "postcode_fake"],
        ["gender", "occupation", "dob"],
        ["first_name", "gender", "dob"],
        ["occupation", "first_and_surname"],
        ["gender", "occupation", "surname"],
        ["first_name", "gender", "full_name"],
        ["gender", "full_name"],
        ["first_name", "postcode_fake"],
        ["occupation", "dob"],
        ["gender", "occupation", "birth_place"],
        ["first_name", "dob"],
        ["gender", "postcode_fake"],
        ["occupation", "surname"],
        ["full_name"],
        ["first_name", "full_name"],
        ["first_name", "gender", "birth_place"],
        ["occupation", "birth_place"],
        ["postcode_fake"],
        ["first_name", "birth_place"],
        ["first_name", "gender", "surname"],
        ["gender", "first_and_surname"],
        ["first_name", "gender", "first_and_surname"],
        ["first_name", "gender", "occupation"],
        ["first_name", "surname"],
        ["first_name", "first_and_surname"],
        ["first_and_surname"],
        ["first_name", "occupation"],
        # ["gender", "surname"],
        # ["surname"],
        # ["gender", "dob"],
        # ["dob"],
        # ["gender", "birth_place"],
        # ["birth_place"],
    ]

    if salt > 1:
        brs = [block_on(x, salting_partitions=salt) for x in brs]
    else:
        brs = [block_on(x) for x in brs]
    return brs


for apply_sort in [True, False]:
    for salt in [1, 10]:
        brs = get_brs(salt)

        settings_dict = {
            "probability_two_random_records_match": 0.0001,
            "link_type": "dedupe_only",
            "blocking_rules_to_generate_predictions": brs,
            "comparisons": [
                jaro_winkler_at_thresholds("first_name"),
                jaro_winkler_at_thresholds("surname"),
                levenshtein_at_thresholds("dob"),
                exact_match("birth_place"),
                levenshtein_at_thresholds("postcode_fake"),
                exact_match("occupation"),
            ],
            "retain_matching_columns": False,
            "retain_intermediate_calculation_columns": False,
        }

        linker = DuckDBLinker(df, settings_dict)

        linker.__apply_sort = apply_sort

        start_time = time.time()
        df_e = linker.predict()
        end_time = time.time()

        print(
            f"Execution time for salt={salt} and "
            f"apply_sort={apply_sort}: {end_time - start_time} seconds"
        )
```

</details>


In the example above I've hacked line 367 of blocking.py to allow __apply_sort
```
    # see https://github.com/duckdb/duckdb/discussions/9710
    # this generates a huge speedup because it triggers parallelisation
    if linker._sql_dialect == "duckdb" and linker.__apply_sort:
        unioned_sql = f"""
        {unioned_sql}
        order by 1
        """
```

```
Execution time for salt=1 and apply_sort=True: 4.619340658187866 seconds
Execution time for salt=10 and apply_sort=True: 11.904595136642456 seconds
Execution time for salt=1 and apply_sort=False: 10.588774919509888 seconds
Execution time for salt=10 and apply_sort=False: 20.690167665481567 seconds
```


Speedup seems to be even greater for a single salted blocking rule.  I would expect them also to be greater the more cores the machine has 

Edit:

This seems to make training u _slower_, despite resulting in all cores being used.  Definitely solvable, but converting to draft for now 

Edit2:  salting for u:

```
import math


def get_salting(max_pairs):
    logged = math.log(max_pairs, 10)
    logged = max(logged - 4, 0)
    return math.ceil(2.5**logged)


for max_pairs in [1, 1e1, 1e2, 1e3, 1e4, 1e5, 1e6, 1e7, 1e8, 2e8, 1e9, 1e10, 1e11]:
    salt = get_salting(max_pairs)
    print(f"{max_pairs:.0e} {salt}")

````