@RobinL 2023-12-12
@NickCrews just for info this is the speedup i mentioned last night