@RobinL 2023-12-12
it's a hunch, but it's because of the way blocking pairs are created and then scored:
- For the first blocking rule, all pairs are generated and then scored
- For the second blocking rules, all pairs are generated, and then pairs created by the first rule are rejected _before_ being scored
- For the third, all pairs are generated, then pairs created by 1st or 2nd are rejected _before_being scored 
etc.

Since scoring is the computationally intensive bit, you want to do this in parallel as much as possible