@RobinL 2023-12-12
<details>
<summary>
time train u
</summary>

```python

for apply_sort in [True]:
    settings_dict = {
        "probability_two_random_records_match": 0.0001,
        "link_type": "dedupe_only",
        "blocking_rules_to_generate_predictions": [],
        "comparisons": [
            jaro_winkler_at_thresholds("first_name"),
            jaro_winkler_at_thresholds("surname"),
            levenshtein_at_thresholds("dob"),
            exact_match("birth_place"),
            levenshtein_at_thresholds("postcode_fake"),
            exact_match("occupation"),
        ],
        "retain_matching_columns": False,
        "retain_intermediate_calculation_columns": False,
    }

    linker = DuckDBLinker(df, settings_dict)

    # linker.__apply_sort = apply_sort

    start_time = time.time()
    df_e = linker.estimate_u_using_random_sampling(1e6)
    end_time = time.time()

    df_e = linker.estimate_u_using_random_sampling(1e6)

    print(
        f"Execution time for  "
        f"apply_sort={apply_sort}: {end_time - start_time} seconds"
    )

```

</details>